[MIG][11.0]hr_contract_multi_jobs:Migrated from 10.0 to 11.0.
