+++ hr_contract_multi_job/__init__.py
@@ -1,0 +1,4 @@
+# Copyright (C) 2014 Savoir-faire Linux. All Rights Reserved.
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+
+from . import models
