+++ hr_contract_multi_job/i18n/sl.po
