+++ hr_contract_multi_job/models/__init__.py
@@ -1,0 +1,6 @@
+# Copyright (C) 2014 Savoir-faire Linux. All Rights Reserved.
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+
+from . import hr_job
+from . import hr_contract
+from . import hr_contract_job
