+++ hr_contract_multi_job/security/ir.model.access.csv
@@ -1,0 +1,2 @@
+id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
+access_hr_contract_job_manager,hr.contract.job.manager,model_hr_contract_job,hr.group_hr_manager,1,1,1,1
