+++ hr_contract_multi_job/tests/test_hr_contract_multi_jobs.py
@@ -1,26 +1,8 @@
-# -*- coding:utf-8 -*-
-##############################################################################
-#
-#    Copyright (C) 2014 Savoir-faire Linux. All Rights Reserved.
-#
-#    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU Affero General Public License as published
-#    by
-#    the Free Software Foundation, either version 3 of the License, or
-#    (at your option) any later version.
-#
-#    This program is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU Affero General Public License for more details.
-#
-#    You should have received a copy of the GNU Affero General Public License
-#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-#
-##############################################################################
+# Copyright (C) 2014 Savoir-faire Linux. All Rights Reserved.
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
-from openerp.tests.common import TransactionCase
-from openerp import exceptions
+from odoo.tests.common import TransactionCase
+from odoo import exceptions
 
 
 class TestContractMultiJob(TransactionCase):
@@ -66,7 +48,7 @@ class TestContractMultiJob(TransactionCase):
         self.contract_id.write({'contract_job_ids':
                                 [(0, 0, {'job_id': self.job_id.id,
                                          'is_main_job': True})]})
-        self.assertTrue(self.contract_id.job_id.id == self.job_id.id)
+        self.assertEqual(self.contract_id.job_id.id, self.job_id.id)
 
     def test_two_contract_jobs_one_main_job(self):
         """
@@ -79,7 +61,7 @@ class TestContractMultiJob(TransactionCase):
                                          'is_main_job': True}),
                                  (0, 0, {'job_id': self.job_2_id.id,
                                          'is_main_job': False})]})
-        self.assertTrue(self.contract_id.job_id.id == self.job_id.id)
+        self.assertEqual(self.contract_id.job_id.id, self.job_id.id)
 
     def test_two_contract_jobs_two_main_job(self):
         """
