+++ hr_contract_multi_job/views/hr_contract_view.xml
@@ -1,27 +1,25 @@
 <?xml version="1.0" encoding="utf-8"?>
-<openerp>
-    <data>
+<odoo>
+        <!-- Inherited Contract view -->
         <record id="hr_contract_view_form" model="ir.ui.view">
             <field name="name">hr.contract.view.form</field>
             <field name="model">hr.contract</field>
             <field name="inherit_id" ref="hr_contract.hr_contract_view_form"/>
             <field name="arch" type="xml">
-
-                <field name="job_id" position="replace">
-                    <field name="job_id" readonly="1"/>
+                <field name="job_id" position="attributes">
+                    <attribute name="readonly">1</attribute>
                 </field>
-
-                <field name="wage" position="after">
-                    <label for="contract_job_ids" string="Job Positions" colspan="2"/>
-                    <field name="contract_job_ids" colspan="2" nolabel="1">
+                <group name="main_info" position="after">
+                    <group name="group_contract" colspan="6">
+                    <separator name="job_position" string="Job Positions"/>
+                    <field name="contract_job_ids" colspan="6" nolabel="1">
                         <tree editable="bottom">
                             <field name="job_id"/>
                             <field name="is_main_job"/>
                         </tree>
                     </field>
-                </field>
-
+                    </group>
+                </group>
            </field>
         </record>
-    </data>
-</openerp>
+</odoo>
