--- hr_contract_multi_jobs/security/ir.model.access.csv
@@ -1,2 +1,0 @@
-id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
-access_hr_contract_job_manager,hr.contract.job.manager,model_hr_contract_job,base.group_hr_manager,1,1,1,1
